+++ src/Lurog.NET/Lurog.NET/ErrorRaisedEventArgs.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Lurog.NET
+{
+    public class ErrorRaisedEventArgs : EventArgs
+    {
+        public int Severity { get; }
+        public string? ErrorCode { get; }
+        public DateTime TimeOccurence { get; }
+        public string? Message { get; }
+    }
+}